BGMフェードインフェードアウトOK
git-svn-id: file:///home/gecchi/repos@2444 9df8dbd1-7660-4e89-b1e7-fa117e7ea0c7

+++ GgafDx9Core/include/jp/ggaf/dx9core/sound/GgafDx9Bgm.h
@@ -21,9 +21,9 @@ public:
     Dix::sp< Dix::OggDecoder> spOggDecoder;
 
     bool _is_fade;
-    float _now_volume;
-    float _target_volume;
-    float _inc_volume;
+    double _now_volume;
+    double _target_volume;
+    double _inc_volume;
 
     GgafDx9Bgm(char* _ogg_name);
